Explicit update of `hwdata` dependency on `releases/v0.23`
The package is a required dependency to build `rocm-dbgapi` on the AMD
platform.  The upstream package is not maintained and packages for spack
`releases/v0.23` only contain versions `0.345` and `0.340` which are old
and fail to build `rocm-dbgapi`.  The `hwdata` for spack v1.0 contains
an update that is not available on the `releases/v0.23` branch which
would be sufficient for successful build of `rocm-dbgapi`
(https://github.com/spack/spack-packages/blob/b893cc3fbf503b5733ed060432b70c11f63c9c4f/repos/spack_repo/builtin/packages/hwdata/package.py#L18).
This commit is unlikely necessary for builds that use spack `v1.0.0`
unless one desires a more recent package version which has been added
here for version `0.396` (the reset of the code is verbatim copy-paste
from upstream package).